Adding tests, renaming State, Fixing category

+++ ewm-service/src/main/java/ru/practicum/events/dto/NewEventDto.java
@@ -27,7 +27,7 @@ public class NewEventDto {
     private String annotation;
 
     @NotNull
-    private int category;
+    private long category;
 
     @NotNull
     @NotBlank
